Add RatioBase Class
I've added the RatioBase class so that I can have a container (map)
that holds multiple types of units in it. I've also added a factory
method for this that is probably not the most necessary thing in
the world but thought it may come in handy later on down the road.